Add reference to WebHandler in WebExceptionHandler
`WebExceptionHandler` are meant to process exceptions that were thrown
while handling the original exchange. Prior to this commit, those
exception handlers would only have a reference to the current exchange
and the exception thrown. With that, the exception handler can only
complete the response (with an error message, for example) or delegate
to the next exception handler.

This commit adds a reference to the original `WebHandler`, so that
`WebExceptionHandler` instances can choose to delegate the error
handling back to the original `WebHandler`; of course, the exchange must
be mutated before that, or the same exception will be thrown.

This allows for more advanced error handling, similar to what's
available in the Servlet world with error pages, i.e. dispatching errors
back into the container for error handling.

Issue: SPR-15812