Add client responder configuration
Prior to this commit, the `RSocketRequester.Builder` would allow to
configure directly annotated handlers for processing server requests.
This lead to a package tangle where the `o.s.messaging.rsocket` would
use classes from `o.s.messaging.rsocket.annotation.support` package.

This commit introduces the `ClientResponder` interface for configuring a
responder on the client requester. This is compatible with future
changes with a parallel implementation of a functional variant for
RSocket handlers.

Closes gh-23170